Seperate context menu from symbol
